+++ IISU/WinCertJobTypeBase.cs
@@ -27,6 +27,7 @@ namespace Keyfactor.Extensions.Orchestrator.WindowsCertStore
     {
         None,
         WinIIS,
-        WinSQL
+        WinSQL,
+        WinLdap
     }
 }